Change event card to teal #A8DADC, buttons to blue #1E88E5, full month name

+++ app.js
@@ -1294,7 +1294,7 @@ function formatLongDate(value) {
   return capitalize(new Intl.DateTimeFormat('sv-SE', {
     weekday: 'long',
     day: 'numeric',
-    month: 'short'
+    month: 'long'
   }).format(date));
 }
